+++ packages/dynamic-bonding-curve/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to the Dynamic Bonding Curve SDK will be documented in this file.
 
+## [1.2.9] - 2025-06-26
+
+### Added
+
+- `TokenUpdateAuthorityOption` enum to have more options for token update authority:
+    - CreatorUpdateAuthority (0)
+    - Immutable (1)
+    - PartnerUpdateAuthority (2)
+    - CreatorUpdateAndMintAuthority (3)
+    - PartnerUpdateAndMintAuthority (4)
+
+### Changed
+
+- Changed `CollectFeeMode` enums from `OnlyQuote` and `Both` to `QuoteToken` and `OutputToken`
+
 ## [1.2.8] - 2025-06-24
 
 ### Added